Remove commentted out code
