+++ test/UserRepository-test.js
@@ -12,7 +12,6 @@ import {
 import {
   userData
 } from '../src/data/users.js';
-console.log(userData)
 
 
 describe('User Repository', () => {
@@ -42,28 +41,7 @@ describe('User Repository', () => {
     expect(userRepo.getAvgStepGoalOfAllUsers()).to.equal(6700);
   });
 
-  it('Should return how much average water consumed for all times', () => {
-    expect(userRepo.waterConsumedAllTime(hydrationData)).to.equal(65);
-  });
 
-  it.only('Should find how many fluid ounces of water consumed each day over the course of a week', () => {
-    expect(userRepo.waterConsumedOverWeek(hydrationData).length).to.equal(7);
-  });
 
-  it('Should find a user/s average sleep quality per day over all time', () => {
-    expect(userRepo.getAvgSleepQualityOfUser(sleepData)).to.equal(3);
-  });
-
-  it('Should find a user/s hours slept each day over the course of a given week', () => {
-    expect(userRepo.getHoursSleptOverWeek(sleepData, 1, "2019/06/15", "2019/06/16", "2019/06/17", "2019/06/18", "2019/06/19", "2019/06/20", "2019/06/21")).to.deep.equal([6.1, 4.1, 8, 10.4, 10.7, 9.3, 7.8]);
-  });
-
-  it('Should find a user/s sleep quality for each day over the course of a given week', () => {
-    expect(userRepo.getSleepQualityForWeek(sleepData, 1, "2019/06/15", "2019/06/16", "2019/06/17", "2019/06/18", "2019/06/19", "2019/06/20", "2019/06/21")).to.deep.equal([2.2, 3.8, 2.6, 3.1, 1.2, 1.2, 4.2]);
-  });
-
-  it('Should find a user/s average minutes active per day over the course of a given week', () => {
-    expect(userRepo.getMinActiveForWeek(activityData, 1, "2019/06/15", "2019/06/16", "2019/06/17", "2019/06/18", "2019/06/19", "2019/06/20", "2019/06/21")).to.equal(171.14285714285714);
-  });
 
 });
